Complete ClientControllerTesting along with revisiting several methods, ensuring correct functionality

+++ ClientDashboard_API/Controllers/ClientController.cs
@@ -1,5 +1,4 @@
-﻿using ClientDashboard_API.Entities;
-using ClientDashboard_API.Interfaces;
+﻿using ClientDashboard_API.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ClientDashboard_API.Controllers
@@ -11,13 +10,13 @@ namespace ClientDashboard_API.Controllers
         /// within their respective block
         /// </summary>
         [HttpGet("{clientName}/currentSession")]
-        public async Task<ActionResult<Client>> GetCurrentClientBlockSession(string clientName)
+        public async Task<ActionResult<int>> GetCurrentClientBlockSessionAsync(string clientName)
         {
             var client = await unitOfWork.ClientRepository.GetClientByNameAsync(clientName);
 
             if (client == null) return NotFound($"{clientName} was not found");
 
-            return Ok(client);
+            return Ok(client.CurrentBlockSession);
         }
 
 
@@ -25,7 +24,7 @@ namespace ClientDashboard_API.Controllers
         /// Client method allowing for the retrieval of all clients, on their last block session
         /// </summary>
         [HttpGet("/onLastSession")]
-        public async Task<ActionResult<List<string>>> GetClientsOnLastBlockSession()
+        public async Task<ActionResult<List<string>>> GetClientsOnLastBlockSessionAsync()
         {
             var clientSessions = await unitOfWork.ClientRepository.GetClientsOnLastSessionAsync();
 
@@ -37,7 +36,7 @@ namespace ClientDashboard_API.Controllers
         /// Client method allowing for the retrieval of all clients 
         /// </summary>
         [HttpGet("/onFirstSession")]
-        public async Task<ActionResult<List<string>>> GetClientsOnFirstBlockSession()
+        public async Task<ActionResult<List<string>>> GetClientsOnFirstBlockSessionAsync()
         {
             var clientSessions = await unitOfWork.ClientRepository.GetClientsOnFirstSessionAsync();
 
@@ -50,8 +49,12 @@ namespace ClientDashboard_API.Controllers
         /// Client method for adding a new Client to the database
         /// </summary>
         [HttpPost]
-        public async Task<IActionResult> AddNewClient(string clientName, int? blockSessions)
+        public async Task<IActionResult> AddNewClientAsync(string clientName, int? blockSessions)
         {
+            var clientExists = await unitOfWork.ClientRepository.CheckIfClientExistsAsync(clientName);
+
+            if (clientExists) return BadRequest($"Client {clientName} already exists in the database");
+
             await unitOfWork.ClientRepository.AddNewClientAsync(clientName, blockSessions);
             if (await unitOfWork.Complete()) return Ok($"Client: {clientName} added");
 
@@ -62,9 +65,10 @@ namespace ClientDashboard_API.Controllers
         /// Client method for removing an existing Client from the database
         /// </summary>
         [HttpDelete]
-        public async Task<IActionResult> RemoveClient(string clientName)
+        public async Task<IActionResult> RemoveClientAsync(string clientName)
         {
             var client = await unitOfWork.ClientRepository.GetClientByNameAsync(clientName);
+            if (client == null) return NotFound($"Client {clientName} not found in the database");
 
             unitOfWork.ClientRepository.RemoveClient(client);
             if (await unitOfWork.Complete()) return Ok($"Client: {clientName} removed");
